scrapper/requirements.txt
pika==1.3.1
aiohttp==3.7.4
beautifulsoup4==4.9.3
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3
